can-j1939: drop rtnetlink for addresses
Addresses & names are assigned on the fly when sockets need it.
The need for keeping a list of local addresses & names still exist
for the transport protocol: we need to know which sessions we need
to contribute to, and which we remain passive.
Since the socket layer deals with promiscuous operation, it became
redundant to have a seperate file for it, only to potentially block
foreign transport protocol sessions. Moreover, it became difficult
to decide on that point.

Signed-off-by: Kurt Van Dijck <dev.kurt@vandijck-laurijssen.be>